Refactor
- Add directories
- Remove unused functions
- Rename files
