+++ src/hittable/sphere.c
@@ -6,7 +6,7 @@
 /*   By: katakagi <katakagi@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/01 12:11:58 by susami            #+#    #+#             */
-/*   Updated: 2023/02/09 21:30:44 by katakagi         ###   ########.fr       */
+/*   Updated: 2023/02/13 13:43:12 by susami           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
