+++ src/utils/degrees_to_radians.c
@@ -1,24 +1,19 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   vec.c                                              :+:      :+:    :+:   */
+/*   degrees_to_radians.c                               :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: katakagi <katakagi@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*   By: susami <susami@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/01/31 14:37:27 by susami            #+#    #+#             */
-/*   Updated: 2023/02/09 21:33:36 by katakagi         ###   ########.fr       */
+/*   Created: 2023/02/13 13:40:50 by susami            #+#    #+#             */
+/*   Updated: 2023/02/13 13:41:11 by susami           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "vec.h"
+#include "minirt.h"
 #include <math.h>
 
-t_vec	vec_new(FLOAT x, FLOAT y, FLOAT z)
+FLOAT	degrees_to_radians(FLOAT degrees)
 {
-	return ((t_vec){.x = x, .y = y, .z = z});
-}
-
-t_point	point_new(FLOAT x, FLOAT y, FLOAT z)
-{
-	return ((t_point){.x = x, .y = y, .z = z});
+	return (degrees * M_PI / 180);
 }
