+++ src/vec/vec_arith.c
