+++ src/vec/vec_utils.c
